Correct layer depends and compat machine
Following feedback from Mark Hatle, corrected the layer dependencies to
include the wrlimnux layer and wr-templates. The wr-templates layer is
only required when a layer itself includes one or more templates.

Also changed the COMPATIBLE_MACHINE definition to include the _rpi
extension and specifically set it to "rpi". The previous definition
could possibly have unintended consequences for other BSPs if this layer
was included but not used in a build.