Remove BSP suffix from common vars
Several variables had an _rpi suffix. Rather than adding an _rpi2
version of each, it is possible to remove the suffix and make the
variables common to all BSPs using this layer.

For the scc & cfg file, we need to ensure only the versions for the
selected machine are used so replace rpi prefix with ${MACHINE}
variable.