Update README with new setup options
As of LTS-21 it is possible to specify custom layers, including BSP
layers, on the setup.sh command-line. It is also then possible to
select a BSP from a custom layer.

NOTE: It is not possible to specify custom templates from those
layers as template validation happens before the layers are added.