Fix TCINDEX configuration warning in RCPL 8
In RCPL 8 the following warning is seen when the kernel configuration
check is performed:

	[INFO]: the following symbols were not found in the active
	configuration:
	     - CONFIG_NET_CLS_TCINDEX

This is due to the TCINDEX module no longer being included in the
kernel source, resulting in the option being removed. For details,
see:

	https://git.kernel.org/pub/scm/linux/kernel/git/stable/stable-queue.git/tree/releases/5.15.100/net-sched-retire-tcindex-classifier.patch

To fix the warning, the option must be removed from the cfg files.